add swagger, refactor auth

+++ src/user/user.service.ts
@@ -1,21 +1,18 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
-import { UpdateUserDto } from './dto/update-user.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { UserEntity } from './entities/user.entity';
 import { Repository } from 'typeorm';
-import bcrypt from 'bcrypt';
 
 @Injectable()
 export class UserService {
   constructor(
     @InjectRepository(UserEntity)
     private readonly userRepository: Repository<UserEntity>
-  ) {}
+  ) { }
 
   async create(createUserDto: CreateUserDto): Promise<UserEntity> {
     const { passwordHash, email } = createUserDto;
-    // const passwordHash = await bcrypt.hash(password, 10)
     const newUser = this.userRepository.create({ email, passwordHash });
     return await this.userRepository.save(newUser);
   }
@@ -24,24 +21,32 @@ export class UserService {
     return await this.userRepository.find();
   }
 
-  async findOne(id: number) {
-    return await this.userRepository.findOneByOrFail({ id })
+  async findOneById(id: number): Promise<UserEntity> {
+    const user = await this.userRepository.findOneBy({ id });
+
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+
+    return user;
   }
 
-  async findOneById(id: number): Promise<UserEntity | null> {
-    return await this.userRepository.findOneBy({ id })
-  }
+  async findOneByEmail(email: string): Promise<UserEntity> {
+    const user = await this.userRepository.findOneBy({ email })
 
-  async findOneByEmail(email: string): Promise<UserEntity | null> {
-    return await this.userRepository.findOneBy({ email })
-  }
+    if (!user) {
+      throw new NotFoundException('User not found')
+    }
 
-  update(id: number, updateUserDto: UpdateUserDto) {
-    return `This action updates a #${id} user`;
+    return user;
   }
 
   async remove(id: number) {
-    const user = await this.findOne(id);
+    const user = await this.findOneById(id);
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+
     await this.userRepository.remove(user);
     return user;
   }
